DFropenode: DFRope.sway is s8
The rope's sway counter is compared against a negative literal --
(s8)self->sway < -DFBARREL_SWAY_LIMIT -- which no unsigned value can ever
satisfy, and it is fed into the node force as (f32)(int)(s8)self->sway. It
oscillates by ++/-- between the two limits, so it is a signed swing, not a
count. Typing it s8 drops the three read casts; DFropenode.o byte-identical.
DFRope.direction stays u8: its own (s8) read cast is inert against the 0/1
values it holds.

+++ src/dlls/objects/373_DFropenode/DFropenode.c
@@ -200,10 +200,10 @@ void DFRope_UpdateSimulation(DFRope* self) {
     partsInit = self->nodes;
     parts = partsInit;
 
-    if ((s8)self->sway < -DFBARREL_SWAY_LIMIT) {
+    if (self->sway < -DFBARREL_SWAY_LIMIT) {
         self->direction = DFBARREL_SWAY_DIR_INCREASING;
     }
-    if ((s8)self->sway > DFBARREL_SWAY_LIMIT) {
+    if (self->sway > DFBARREL_SWAY_LIMIT) {
         self->direction = DFBARREL_SWAY_DIR_DECREASING;
     }
     if ((s8)self->direction == DFBARREL_SWAY_DIR_DECREASING) {
@@ -217,7 +217,7 @@ void DFRope_UpdateSimulation(DFRope* self) {
     {
         f32 rate = lbl_803E4DF8;
         for (; i < self->count - 1; i++) {
-            partIter->force[0] = partIter->force[0] + rate * (f32)(int)(s8)self->sway;
+            partIter->force[0] = partIter->force[0] + rate * (f32)(int)self->sway;
             partIter++;
         }
     }
